+++ src/component/SearchBar.js
@@ -1,48 +1,17 @@
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useEffect, useContext } from 'react';
 import { useHistory } from 'react-router';
 import Context from '../context/Context';
 
-function SearchBar() {
-  const [filterRadio, setFilterRadio] = useState('s');
-  const [filterText, setFilterText] = useState('');
+export default function SearchBar() {
   const history = useHistory();
-  const { setFilteredItem } = useContext(Context);
-
-  async function fetchFood() {
-    let endPoint = `https://www.themealdb.com/api/json/v1/1/search.php?${filterRadio}=${filterText}`;
-    if (filterRadio === 'i') {
-      endPoint = `https://www.themealdb.com/api/json/v1/1/filter.php?i=${filterText}`;
-    }
-    const request = await fetch(endPoint);
-    const response = await request.json();
-    const data = response.meals;
-    if (data !== null) {
-      setFilteredItem(data);
-      if (data.length === 1) {
-        history.push(`/comidas/${data[0].idMeal}`);
-      }
-    } else {
-      global.alert('Sinto muito, não encontramos nenhuma receita para esses filtros.');
-    }
-  }
-
-  async function fetchDrink() {
-    let endPoint = `https://www.thecocktaildb.com/api/json/v1/1/search.php?${filterRadio}=${filterText}`;
-    if (filterRadio === 'i') {
-      endPoint = `https://www.thecocktaildb.com/api/json/v1/1/filter.php?i=${filterText}`;
-    }
-    const request = await fetch(endPoint);
-    const response = await request.json();
-    const data = response.drinks;
-    if (data !== null) {
-      setFilteredItem(data);
-      if (data.length === 1) {
-        history.push(`/bebidas/${data[0].idDrink}`);
-      }
-    } else {
-      global.alert('Sinto muito, não encontramos nenhuma receita para esses filtros.');
-    }
-  }
+  const { filterRadio,
+    setFilterRadio,
+    filterText,
+    setFilterText,
+    fetchFood,
+    fetchDrink,
+    filteredItem,
+  } = useContext(Context);
 
   useEffect(() => {
     if (filterRadio === 'f' && filterText.length > 1) {
@@ -50,9 +19,13 @@ function SearchBar() {
     }
   }, [filterText]);
 
-  // useEffect(() => {
-  //   console.log(filteredItem);
-  // }, [filteredItem]);
+  useEffect(() => {
+    if (filteredItem.length === 1 && history.location.pathname === '/comidas') {
+      history.push(`/comidas/${filteredItem[0].idMeal}`);
+    } if (filteredItem.length === 1 && history.location.pathname === '/bebidas') {
+      history.push(`/bebidas/${filteredItem[0].idDrink}`);
+    }
+  }, [filteredItem]);
 
   return (
     <form id="form">
@@ -105,38 +78,3 @@ function SearchBar() {
     </form>
   );
 }
-
-export default SearchBar;
-
-// import React from 'react';
-// import RadialInput from '../mini-components/RadialInput';
-// import Button from '../mini-components/Button';
-
-// function SearchBar() {
-//   return (
-//     <div>
-//       <input data-testid="search-input" type="text" placeholder="DIGITE SUA COMIDA" />
-//       <RadialInput
-//         dataTest="ingredient-search-radio"
-//         radialId="ingredient"
-//         radialName="search-checkbox"
-//         radialText="Ingrediente"
-//       />
-//       <RadialInput
-//         dataTest="name-search-radio"
-//         radialId="food-name"
-//         radialName="search-checkbox"
-//         radialText="Nome"
-//       />
-//       <RadialInput
-//         dataTest="first-letter-search-radio"
-//         radialId="food-first-letter"
-//         radialName="search-checkbox"
-//         radialText="Primeira Letra"
-//       />
-//       <Button btnText="Buscar" data-testid="exec-search-btn" />
-//     </div>
-//   );
-// }
-
-// export default SearchBar;
